+++ src/mock/movie-mock.js
@@ -1,12 +1,10 @@
 import {getRandomPositiveInteger, getRandomPositiveFloat, getRandomArray, getRandomElement, getRandomDate} from '../utils/common.js';
 import {arrayMovieInfo, workingGroup, ageRestrictionsArray, countryArray, releaseArray, descriptionTextArray} from './../data.js';
-import {generateComment} from './comment.js';
+import {comments} from '../main.js';
 import {nanoid} from 'nanoid';
 import dayjs from 'dayjs';
 import duration from 'dayjs/plugin/duration';
 dayjs.extend(duration);
-
-const COMMENT_COUNT = 10;
 
 const getRandomDuration = () => (
   dayjs.duration({
@@ -15,19 +13,13 @@ const getRandomDuration = () => (
   }).format('H m')
 );
 
-const commentsArray = new Array(COMMENT_COUNT).fill().map(generateComment);
-const commentsIdArray = Array.from(commentsArray).map((el) => el.aboutFilm);
-
-const getCountComments = (comments, id) => {
-  const result = comments.filter((i) => i === id).length;
-  return result;
-};
-
 const generateMovie = () => {
   const indexMovieTitle = getRandomPositiveInteger(1, Object.keys(arrayMovieInfo).length);
-
+  const commentsIdArray = comments.filter((commentElement) => commentElement.aboutFilm === arrayMovieInfo[indexMovieTitle].filmId);
   const movie = {
     id: nanoid(),
+    comments: commentsIdArray,
+
     filmId: arrayMovieInfo[indexMovieTitle].filmId,
     originalName: arrayMovieInfo[indexMovieTitle].original,
     title: arrayMovieInfo[indexMovieTitle].title,
@@ -42,14 +34,11 @@ const generateMovie = () => {
     actors: getRandomArray(workingGroup, 4),
     country: getRandomElement(countryArray),
     ageRestrictions: getRandomElement(ageRestrictionsArray),
-    countComments: getCountComments(commentsIdArray, arrayMovieInfo[indexMovieTitle].filmId),
     isFavorite: Boolean(getRandomPositiveInteger(0, 1)),
     isHistory: Boolean(getRandomPositiveInteger(0, 1)),
     isWatchlist: Boolean(getRandomPositiveInteger(0, 1)),
-    isChoosenEmojiForComment: '',
-    commentText: '',
   };
   return movie;
 };
 
-export { generateMovie, commentsArray};
+export {generateMovie};
