+++ src/view/popup-movie-info.js
@@ -1,6 +1,6 @@
 import SmartView from './smart.js';
-import {emojiArray} from '../data.js';
 import dayjs from 'dayjs';
+import CommentsView from '../view/movie-view';
 
 const createGenresList = (array) => {
   let code = '';
@@ -11,62 +11,11 @@ const createGenresList = (array) => {
   return code;
 };
 
-const createCommentsTemplate = (array) => {
-  let code = '';
-  array.forEach((comment) => {
-    const { emoji, text, date, autor } = comment;
-    const oneComment = `
-      <li class="film-details__comment">
-        <span class="film-details__comment-emoji">
-          <img src="./images/emoji/${emoji}.png" width="55" height="55" alt="emoji-smile">
-        </span>
-        <div>
-          <p class="film-details__comment-text">${text}</p>
-          <p class="film-details__comment-info">
-            <span class="film-details__comment-author">${autor}</span>
-            <span class="film-details__comment-day">${date}</span>
-            <button class="film-details__comment-delete">Delete</button>
-          </p>
-        </div>
-      </li>`;
-    code = code + oneComment;
-  });
-  return `<ul class="film-details__comments-list">${code}</ul>`;
-};
-
-const createNewCommentContainer = (choosenEmoji, commentText) => {
-  const createEmojiSelectionTemplate = () => (
-    emojiArray.map((emojiItem) => `
-      <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-${emojiItem}" value="${emojiItem}" ${choosenEmoji === emojiItem ? 'checked' : ''}>
-      <label class="film-details__emoji-label" for="emoji-${emojiItem}" data-value="${emojiItem}">
-        <img src="./images/emoji/${emojiItem}.png" width="30" height="30" alt="emoji">
-      </label>
-    `).join('')
-  );
-
-  const emojiSelectionTemplate = createEmojiSelectionTemplate();
-  const emojiSrc = (choosenEmoji) ? `style="background-image: url('/images/emoji/${choosenEmoji}.png'); background-size: contain;"` : '';
-  return (`
-    <div class="film-details__new-comment">
-      <div class="film-details__add-emoji-label" ${emojiSrc}></div>
-
-      <label class="film-details__comment-label">
-        <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment">${commentText}</textarea>
-      </label>
-
-      <div class="film-details__emoji-list">
-        ${emojiSelectionTemplate}
-      </div>
-    </div>
-  `);
-};
-
-const createPopupMovieInfo = (movieData, commentsData) => {
-  const {originalName, title, rating, release, duration, genres, poster, description, isFavorite, isHistory, isWatchlist, directors, writers, actors, country, ageRestrictions, isComments, countComments, isChoosenEmojiForComment, commentText} = movieData;
+const createPopupMovieInfo = (movieData) => {
+  const {originalName, title, rating, release, duration, genres, poster, description, isFavorite, isHistory, isWatchlist, directors, writers, actors, country, ageRestrictions, comments} = movieData;
   const releaseWithFormat = dayjs(release).format('D MMMM YYYY');
   const genresList = createGenresList(genres);
-  const commentsTemplate = createCommentsTemplate(commentsData);
-  const getNewComment = createNewCommentContainer(isChoosenEmojiForComment, commentText);
+  const commentsBlock = new CommentsView();
 
   return (
     `<section class="film-details">
@@ -135,12 +84,10 @@ const createPopupMovieInfo = (movieData, commentsData) => {
             <button type="button" class="film-details__control-button--favorite ${(isFavorite) ? 'film-details__control-button film-details__control-button--active' : 'film-details__control-button'}" id="favorite" name="favorite">Add to favorites</button>
           </section>
         </div>
-
         <div class="film-details__bottom-container">
           <section class="film-details__comments-wrap">
-            <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${countComments}</span></h3>
-            ${isComments ? commentsTemplate : ''}
-            ${getNewComment}
+            <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${comments.length}</span></h3>
+            ${comments.length > 0 ? commentsBlock : ''}
           </section>
         </div>
       </form>
@@ -149,23 +96,18 @@ const createPopupMovieInfo = (movieData, commentsData) => {
 };
 
 export default class MoviePopup extends SmartView {
-  constructor(movie, comments) {
+  constructor(movie) {
     super();
-    this._data = MoviePopup.parseMovieToData(movie, comments);
-    this._comments = comments;
-
-    this._formSubmitHandler = this._formSubmitHandler.bind(this);
+    this._data = MoviePopup.parseMovieToData(movie);
     this._closeClickHandler = this._closeClickHandler.bind(this);
     this._watchlistClickPopupHandler = this._watchlistClickPopupHandler.bind(this);
     this._favoriteClickPopupHandler = this._favoriteClickPopupHandler.bind(this);
     this._historyClickPopupHandler = this._historyClickPopupHandler.bind(this);
-    this._emojiClickHandler = this._emojiClickHandler.bind(this);
-    this._commentTextareaHandler = this._commentTextareaHandler.bind(this);
     this._setInnerHandlers();
   }
 
   getTemplate() {
-    return createPopupMovieInfo(this._data, this._comments);
+    return createPopupMovieInfo(this._data);
   }
 
   setCloseClickHandler(callback) {
@@ -179,11 +121,6 @@ export default class MoviePopup extends SmartView {
   }
 
   _setInnerHandlers(){
-    const emojis = this.getElement().querySelectorAll('.film-details__emoji-label');
-    emojis.forEach((emoji) => {
-      emoji.addEventListener('click', this._emojiClickHandler);
-    });
-    this.getElement().querySelector('.film-details__comment-input').addEventListener('input', this._commentTextareaHandler);
     this.getElement().querySelector('.film-details__close-btn').addEventListener('click', this._closeClickHandler);
   }
 
@@ -222,46 +159,12 @@ export default class MoviePopup extends SmartView {
     this._callback.historyClick();
   }
 
-  _emojiClickHandler(e) {
-    e.preventDefault();
-    const value = e.target.parentElement.dataset.value;
-    this.updateData({
-      isChoosenEmojiForComment: value,
-    }, false);
-  }
-
-  _formSubmitHandler(e) {
-    e.preventDefault();
-    this._callback.formSubmit(MoviePopup.parseDataToMovie(this._data));
-  }
-
-  _commentTextareaHandler(e) {
-    e.preventDefault();
-    this.updateData({
-      commentText: e.target.value,
-    }, true);
-    if(e.target.value !== ''){
-      this.setFormSubmitHandler(this._callback.formSubmit);
-    }
-  }
-
-  setFormSubmitHandler(callback) {
-    this._callback.formSubmit = callback;
-    //навесить обработчик
-    document.addEventListener('keydown', (e) => {
-      if((e.code === 'Enter') && e.ctrlKey) {
-        //отправка данных в модель
-      }
-    });
-  }
-
-  static parseMovieToData(movie, comments) { // информация в состояние (когда редактируем)
+  static parseMovieToData(movie) { // информация в состояние (когда редактируем)
     return Object.assign(
       {},
       movie,
       {
-        isComments: comments.length > 0,
-        isChoosenEmoji: movie.isChoosenEmojiForComment,
+        isComments: movie.countComments > 0,
       },
     );
   }
@@ -272,11 +175,6 @@ export default class MoviePopup extends SmartView {
       data.countComments = 0;
     }
 
-    if(!data.isChoosenEmojiForComment) {
-      data.isChoosenEmojiForComment = '';
-    }
-
-    delete data.isChoosenEmojiForComment;
     delete data.isComments;
     return data;
   }
